# Fix backports of std lib modules like `dataclasses` not working with dependency inference (Cherry-pick of #12818)