--- a/src/python/pants/backend/python/dependency_inference/python_stdlib/BUILD
+++ b/src/python/pants/backend/python/dependency_inference/python_stdlib/BUILD
@@ -1,4 +1,0 @@
-# Copyright 2020 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_library()