… src/python/pants/backend/python/dependency_inference/python_stdlib/py27.py
… src/python/pants/backend/python/dependency_inference/python_stdlib/py35.py
… src/python/pants/backend/python/dependency_inference/python_stdlib/py36.py
… src/python/pants/backend/python/dependency_inference/python_stdlib/py37.py
… src/python/pants/backend/python/dependency_inference/python_stdlib/py38.py