Omitting these packages would break backport libraries (such as dataclasses). Performance impact is negligible since the dependency graph has already been gathered.

Benchmarks using 
`$ hyperfine --warmup=1 --runs=10 './pants --no-remote-cache-{read,write} --no-pantsd dependencies ::'`

Before these changes (omitting stdlib dependencies):
```bash
Benchmark #1: ./pants --no-remote-cache-{read,write} --no-pantsd dependencies ::
  Time (mean ± σ):      8.852 s ±  0.105 s    [User: 8.612 s, System: 5.314 s]
  Range (min … max):    8.702 s …  9.107 s    10 runs
```

After changes (with stdlib dependency inference)
```bash
Benchmark #1: ./pants --no-remote-cache-{read,write} --no-pantsd dependencies ::
  Time (mean ± σ):      8.854 s ±  0.123 s    [User: 8.619 s, System: 5.333 s]
  Range (min … max):    8.682 s …  9.111 s    10 runs
```

[ci skip-rust]
[ci skip-build-wheels]